Add random string generator in utils.js
A function `generateRandom(length)` has been added in utils.js, which generates a random string of the specified length. It uses a predefined set of characters and is useful for generating unique identifiers or tokens. Adjustments have also been made in landing-page.js for incorporating the use of this function.