Refine data migration with transformation functions and logs
In the migration script, added transformation functions for 'products', 'batches', and 'audit' data during the transition to LightDB. Introduced waiting function to ensure data transfer completion and termination of both the LightDBEnclave and the server. Also, now the progress of data migration is traced with console logs which output information about each completed stage of the process.