Add environment variable validation script
This commit introduces a new script (000_checkEnvironmentVariables.js) which handles validation of environment variables. This ensures that each variable is defined and contains an expected value, which is crucial during the system setup process. Furthermore, it includes an additional check on 'BDNS_ROOT_HOSTS' and 'SSO_SECRETS_ENCRYPTION_KEY' to ensure proper configuration.