Store initial URL in sessionStorage before logout
The initial URL of the session is now saved in sessionStorage before triggering a logout process. This change has been implemented in Timer.js and LogoutController.js scripts, and applied to all instances of logout across the application. The URL storing will facilitate redirecting user back to the initial page after a potential re-login.