Refactor creation and handling of decentralized identifiers (DIDs)
The method for creating and handling DIDs in landing-page.js was simplified. Instead of trying to resolve DIDs in a loop until a valid one is found, it now tries to resolve the main DID and creates a new one if necessary. The PermissionsWatcher service was also updated to retrieve the EPI group from the server and throw an error if it is not returned as expected.