Add check for already migrated FixedURL FS adapter
This commit introduces a check in the migration script to see if the FixedURL FS adapter has already been migrated. The check will prevent unnecessary migrations and errors in case the script is run multiple times. The path of the migrated file has been abstracted to a constant for consistency and clarity.